Move nasp into separate package
