+++ experimental/python/Makefile
@@ -1,7 +1,7 @@
 include ../../common.mk
 
 dynamic-lib:
-	go build -o nasp.so -buildmode=c-shared nasp.go
+	go build -o nasp/nasp.so -buildmode=c-shared nasp/nasp.go
 
 static-lib:
-	go build -o nasp.a -buildmode=c-archive nasp.go
+	go build -o nasp/nasp.a -buildmode=c-archive nasp/nasp.go
